Better checks for aws creds, and centralised those checks in the load_and_validate_platform_config function
